requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.4.0
psycopg2-binary>=2.9.9
yfinance>=0.2.54
pandas>=2.2.3
